+++ src/http-client/index.ts
@@ -12,12 +12,12 @@ declare const __USE_AXIOS__: boolean;
 // Use the variable for the runtime check
 // eslint-disable-next-line no-lonely-if
 if (__USE_AXIOS__) {
-  // eslint-disable-next-line global-require
+  // eslint-disable-next-line global-require, prefer-import/prefer-import-over-require
   const axiosModule = require('./axios-client');
   httpClient = axiosModule.axiosClient;
   create = axiosModule.create;
 } else {
-  // eslint-disable-next-line global-require
+  // eslint-disable-next-line global-require, prefer-import/prefer-import-over-require
   const fetchModule = require('./fetch-client');
   httpClient = fetchModule.fetchClient;
   create = fetchModule.create;